+++ server/Cargo.toml
@@ -29,7 +29,6 @@ kona-providers-alloy = { workspace = true }
 # Alloy
 alloy-primitives = { workspace = true, features = ["map", "serde"] }
 alloy-provider = { workspace = true, features = ["reqwest"] }
-alloy-transport-http = { workspace = true }
 alloy-transport = { workspace = true }
 alloy-rpc-client = { workspace = true }
 alloy-json-rpc = { workspace = true }
